requirements-asm-patterns.txt
capstone==5.0.5
keystone-engine==0.9.2
lief==0.16.4
mypy==1.15.0
mypy-extensions==1.0.0
typing_extensions==4.12.2
